backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
celery[redis]==5.4.0
redis==5.2.1
aioredis==2.0.1
boto3==1.36.2
minio==7.2.12
pymupdf==1.25.3
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.1.0
openpyxl==3.1.5
weasyprint==63.1
langchain==0.3.14
langchain-core==0.3.30
langchain-community==0.3.14
langchain-openai==0.3.1
langgraph==0.2.65
openai==1.59.9
anthropic==0.42.0
google-generativeai==0.8.4
httpx==0.28.1
pgvector==0.3.6
python-magic==0.4.27
sse-starlette==2.2.1
slowapi==0.1.9
bcrypt==4.2.1
emails==0.6
jinja2==3.1.5
aiofiles==24.1.0
numpy==2.2.2
tiktoken==0.8.0
